Supports recursive domain lookup in DB
Previously, when only a single subdomain is supported, which creates a
bug. For example, user query massbrowser.cs.umass.edu and first the
program look for massbrowser.cs.umass.edu if it could not find it, it
will try cs.umass.edu. Howevre, if it couldn't find cs.umass.edu, it
assumes no rules apply. This is wrong because umass.edu may exist as a
rule.

Now it supports to query it recursively until no match can be found.

Many thanks to @srxzr for spending time helping me debug and teaching
me the best workflow for MassBrowser development.